+++ CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-### Added
+## [2.2.1] - 2019-06-19
+### Added
 - New comparison methods on some domain types.
 
 ### Changed
-- When the session is decoded but null or empty return null.
+- When an auth token is decoded into a session but is null or empty now return null.
 
 ### Fixed
 - Awaited socket callback tasks are now canceled when the socket adapter is closed and cleared.